refactor(cli): extract sandbox rebuild action
